requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astroid==2.7.3
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
dataclasses==0.6
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.3
Django==2.1.15
entrypoints==0.3
filelock==3.3.0
Flask==2.0.2
Flask-DebugToolbar==0.11.0
flask-heroku==0.1.9
gitdb==4.0.7
GitPython==3.1.24
gotrue==0.2.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pbr==5.6.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
plotly==5.3.1
pluggy==1.0.0
postgrest-py==0.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
py==1.11.0
pyarrow==5.0.0
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pylint==2.10.2
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
rjsmin==1.1.0
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
sodapy==2.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.27
stevedore==3.4.0
streamlit==0.89.0
supabase-py==0.0.2
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
venv-run==0.1.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==3.0.0
xlwings==0.25.1
yfinance==0.1.63
